models: Added ContainerManagerConfig.
The new ContainerManagerConfig model is meant to store configuration
data used when creating the upcomming ContainerManager instance. When a
cluster has a ContainerManager associated with it the controlling code
should match the Cluster.container_manager to a
ContainerManagerConfig.name. The resulting ContainerManager can then be
used to work with proper external ContainerManager.

Note that not all Clusters need a ContainerManager. The Cluster is not
associated with a ContainerManager if it's container_manager field is
empty.